pyproject.toml
[tool.poetry]
name = "discordjingleonjoin"
version = "0.1.0"
description = ""
authors = ["Simon Goričar <simon.peter.goricar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {version = "^1.7.3", extras = ["voice"]}
toml = "^0.10.2"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
